chore: refactor TLS configuration to avoid circular dependencies
