+++ server/config/tls_options.go
@@ -1,4 +1,4 @@
-package server
+package config
 
 import (
 	"crypto/tls"
